% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries_get.R
\name{boundaries_get}
\alias{boundaries_get}
\title{Retrieve Geographical Boundaries}
\usage{
boundaries_get(boundary, year = NULL, detail_level = "BUC")
}
\arguments{
\item{boundary}{a string containing the acronym of name of the boundary to be
downloaded. Accepted values are: CAUTH, CED, CTRY, CTY, CTYUA, LAD, LPA, LSOA, MCTY, MSOA, NHS, OA, PAR, PARNCP, RGN, UTLA, WD}

\item{year}{a number containing the year when the boundary was created.}

\item{detail_level}{a string defining the level of detail in the geometry.
(this affects the download size).
Accepted values are: BFC, BFE, BGC, BSC, BUC. Each value
corresponds to:
\itemize{
\item \emph{Full Extent (BFE)} – Full resolution boundaries go to the Extent of the Realm (Low Water Mark) and are the most detailed of the boundaries.
\item \emph{Full Clipped (BFC)} – Full resolution boundaries that are clipped to the coastline (Mean High Water mark).
\item \emph{Generalised Clipped (BGC)} - Generalised to 20m and clipped to the coastline (Mean High Water mark) and more generalised than the BFE boundaries.
\item \emph{Super Generalised Clipped (BSC) (200m)} – Generalised to 200m and clipped to the coastline (Mean High Water mark).
\item \emph{Ultra Generalised Clipped (BUC) (500m)} – Generalised to 500m and clipped to the coastline (Mean High Water mark).
\item \emph{Grid, Extent (BGE)} - Grid formed of equally sized cells which extend beyond the coastline.
\item \emph{Generalised, Grid (BGG)} - Generalised 50m grid squares.
}

For a detailed description of how these boundaries were created refer to
\href{https://www.ons.gov.uk/methodology/geography/geographicalproducts/digitalboundaries}{Digital boundaries}}
}
\value{
a \code{sf} object with the selected boundaries.
}
\description{
Queries \href{https://geoportal.statistics.gov.uk/}{ONS' Geoportal} endpoints and
retrieves the requested geographical boundaries in the form of a sf object.
}
\examples{
countries_2023 <- boundaries_get("CTRY", 2023, "BUC")

class(countries_2023)

}
