% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries_select.R
\name{boundaries_select}
\alias{boundaries_select}
\title{Interactively Retrieve Geographical Boundaries}
\usage{
boundaries_select()
}
\value{
a \code{sf} object with the selected combination of boundary, year and
detail level.
}
\description{
A Command Line Interface that constructs a valid combination of geometry
type, year and detail level and downloads the selected boundary by calling
\code{boundaries_get()}.
}
