% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boundaries.R
\name{get_boundaries}
\alias{get_boundaries}
\title{Retrieve Geographical Boundary}
\usage{
get_boundaries(boundary, year, resolution = "BUC")
}
\arguments{
\item{boundary}{a string containing...}

\item{year}{a number containing...}

\item{resolution}{a string containing...}
}
\value{
a sf object
}
\description{
Queries \href{https://geoportal.statistics.gov.uk/}{ONS' Geoportal} endpoints and
retrieves the requested geographical boundaries in the form of a sf object.
}
