% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries_select.R
\name{boundaries_select}
\alias{boundaries_select}
\title{Interactively Retrieve Geographical Boundaries}
\usage{
boundaries_select()
}
\value{
a \code{sf} object with the selected boundaries.
}
\description{
Constructs a valid combination of geometry type, year and detail level and
downloads the selected boundary.
}
